exported following functions
